Use device memory for kv-test AIV I/O
