+++ ucm/transport/kv/kv-test/include/kv_test/kv_test_config_helpers.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "kv_test/kv_test_types.h"
+
+namespace UC::KVTest {
+
+bool IsFakeBackendMode(const KvTestConfig& config);
+bool IsAivProviderMode(const KvTestConfig& config);
+void MaybePrepareFakeBackend(KvTestConfig& config);
+
+}  // namespace UC::KVTest
